requirements.txt
nltk==3.3
opencv_python==3.4.3.18
pygame==1.9.4
h5py==2.8.0
wget==3.2
scipy==1.1.0
numpy==1.15.2
matplotlib==3.0.0
requests==2.19.1
Pillow==5.2.0
beautifulsoup4==4.6.3
seaborn==0.9.0
